much problems, about to swap x and y for good hopefully

+++ includes/cub3d.h
@@ -107,7 +107,7 @@ typedef	struct	s_params
 	int		sqx;
 	int		sqy;
 	int		scrdist;
-	int		sizeconst;
+	float		sizeconst;
 }		t_param;
 
 int	ft_atoi(const char *str);
